pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clonebox"
version = "1.1.5"
description = "Clone your workstation environment to an isolated VM with selective apps, paths and services"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
authors = [
    {name = "CloneBox Team"}
]
keywords = ["vm", "virtualization", "libvirt", "clone", "workstation", "qemu", "kvm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

dependencies = [
    "libvirt-python>=9.0.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.0.0",
    "pytest-asyncio",
    "fastapi>=0.100.0",
]
dashboard = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]

[project.scripts]
clonebox = "clonebox.cli:main"

[project.urls]
Homepage = "https://github.com/wronai/clonebox"
Repository = "https://github.com/wronai/clonebox"
Issues = "https://github.com/wronai/clonebox/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clonebox = ["templates/**/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: end-to-end tests requiring libvirt/KVM (deselect with '-m \"not e2e\"')",
    "slow: slow tests (deselect with '-m \"not slow\"')",
    "integration: integration tests",
    "container: container workflow tests requiring podman/docker",
    "requires_kvm: tests requiring /dev/kvm access",
]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--cov=src/clonebox",
    "--cov-report=term-missing",
    "--cov-fail-under=30",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
omit = [
    "src/clonebox/container.py",
    "src/clonebox/cli.py",
]
